# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/role/add": {
            "post": {
                "description": "添加角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role 角色管理"
                ],
                "summary": "添加角色",
                "operationId": "RoleAdd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "角色信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.RoleAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/role/all": {
            "get": {
                "description": "获取所有角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role 角色管理"
                ],
                "summary": "获取所有角色",
                "operationId": "RoleAll",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.PageResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/resp.RoleSimpleResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/role/change": {
            "post": {
                "description": "角色状态修改",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role 角色管理"
                ],
                "summary": "角色状态修改",
                "operationId": "RoleChange",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "角色ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.IdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/role/delete": {
            "post": {
                "description": "删除角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role 角色管理"
                ],
                "summary": "删除角色",
                "operationId": "RoleDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "角色ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.IdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/role/detail/{id}": {
            "get": {
                "description": "获取角色详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role 角色管理"
                ],
                "summary": "获取角色详情",
                "operationId": "RoleDetail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "角色ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.IdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.RoleResp"
                        }
                    }
                }
            }
        },
        "/auth/role/edit": {
            "post": {
                "description": "编辑角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role 角色管理"
                ],
                "summary": "编辑角色",
                "operationId": "RoleEdit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "角色信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.RoleEditReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/role/list": {
            "get": {
                "description": "获取角色列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role 角色管理"
                ],
                "summary": "获取角色列表",
                "operationId": "RoleList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码",
                        "name": "pageNo",
                        "in": "query"
                    },
                    {
                        "maximum": 60,
                        "type": "integer",
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.PageResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/resp.RoleResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "req.IdReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                }
            }
        },
...
        "req.RoleAddReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "is_admin": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "is_disable": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 125
                },
                "remark": {
                    "type": "string",
                    "maxLength": 255
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
        "req.RoleEditReq": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_admin": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "is_disable": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 125
                },
                "remark": {
                    "type": "string",
                    "maxLength": 255
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
...
        "resp.RoleResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/types.DateTime"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "is_disable": {
                    "description": "是否禁用: [0=否, 1=是]",
                    "type": "integer"
                },
                "member": {
                    "description": "成员数量",
                    "type": "integer"
                },
                "menus": {
                    "description": "关联菜单",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "remark": {
                    "description": "角色备注",
                    "type": "string"
                },
                "sort": {
                    "description": "角色排序",
                    "type": "integer"
                },
                "updated_at": {
                    "$ref": "#/definitions/types.DateTime"
                }
            }
        },
        "resp.RoleSimpleResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/types.DateTime"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/types.DateTime"
                }
            }
        },
...
        "response.PageResp": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "总数",
                    "type": "integer"
                },
                "lists": {
                    "description": "数据"
                },
                "pageNo": {
                    "description": "页No",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页Size",
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.DateTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        }
    }
}